refac : #8 MultiUseReturnCompleteScreen 함수 분리 및 주석 추가
